Use boto for S3-like storage depots
Libcloud's API does not provide a way to get progress of uploads without causing the entire contents of a file to be loaded into memory. Libcloud is still used as a backup for non-S3-like APIs. However, progress will appear stalled for any depot that uses libcloud (local, azure, etc).

Also, boto provides a way to have resumable downloads. And if google storage is used, resumable uploads as well.